Separate desktop-specific and shared markup component code
