+++ shared/common-adapters/index.js
@@ -23,7 +23,7 @@ export {default as Icon} from './icon'
 export {default as Input} from './input'
 export {default as LinkWithIcon} from './link-with-icon'
 export {default as ListItem} from './list-item'
-export {default as Markdown} from './markdown.js'
+export {default as Markdown} from './markdown'
 export {default as MultiAvatar} from './multi-avatar.js'
 export {default as Meta} from './meta'
 export {default as PlatformIcon} from './platform-icon'
